// src/types/ui/input.ts
export type InputVariants = 'filled' | 'outlined' | 'standard' | 'borderless';
export type InputTypes =
  | 'text'
  | 'password'
  | 'email'
  | 'number'
  | 'telephone'
  | 'url'
  | 'checkbox'
  | 'date'
  | 'file'
  | 'dropdown'
  | 'amount'
  | 'alphaNumeric'
  | 'range-number';

export type InputScale =
  | 'base'
  | 'success'
  | 'warning'
  | 'danger'
  | 'primary'
  | 'disabled'
  | 'secondary';
export type InputOrientation = 'vertical' | 'horizontal';
export type InputSizes = 'sm' | 'md' | 'lg' | 'fullWidth';